renamed all functions

+++ scripts/ubuntu.sh
@@ -2,15 +2,22 @@
 
 set -e
 
-basic () {
-sudo apt-get update -y # && sudo apt-get upgrade -y
-# sudo apt  install -y tree tmux nano unzip vim wget git net-tools bind9-utils bridge-utils bash-completion zsh zsh-completion htop jq
-sudo apt-get install -y tree tmux nano unzip vim wget git net-tools zsh htop jq ca-certificates curl gnupg lsb-release
-#sudo apt install -y apt-file tasksel
+install_packages() {
+if [[ ! -f $(which apt-get) ]]
+then
+  sudo apt-get update -y # && sudo apt-get upgrade -y
+  sudo apt-get install -y tree tmux nano unzip vim wget git net-tools zsh htop jq ca-certificates curl gnupg lsb-release
+fi
 
+if [[ ! -f $(which yum) ]]
+then
+  sudo yum update --assumeyes;
+  sudo yum install -y curl tree tmux nano unzip vim wget git net-tools bash-completion zsh zsh-completion bind-utils bridge-utils jq
+  # sudo amazon-linux-extras install epel docker -y; sudo usermod -a -G docker ec2-user
+fi
 }
 
-dev () {
+install_dev() {
 if [[ ! -f $(which nvm) ]]
 then
 # https://github.com/nvm-sh/nvm/#installing-and-updating
@@ -42,7 +49,7 @@ fi
 go version
 }
 
-docker () {
+install_docker() {
 
 if [[ ! -f $(which docker) ]]
 then
@@ -66,7 +73,7 @@ docker-compose --version
 }
 
 
-git () {
+install_git() {
 if [[ ! -f  ~/.gitconfig ]]
 then
 set -x
@@ -74,8 +81,7 @@ curl -o ~/.gitconfig https://raw.githubusercontent.com/amitkarpe/setup/main/dot/
 curl -o ~/.gitignore_global https://raw.githubusercontent.com/amitkarpe/setup/main/dot/.gitignore_global
 
 export PAGER=''
-sleep 2
-git config --global --list
+# git config --global --list
 cat ~/.gitconfig
 set +x
 fi
@@ -84,10 +90,10 @@ fi
 
 main () {
   sleep 2
-  basic
-  dev
-  docker
-  git
+  install_packages
+  install_dev
+  install_docker
+  install_git
 }
 
 main
